requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
pandas==2.2.1
numpy==1.26.4
ta==0.11.0
